Partial auth implementation
Integrated SQLc and Goose for postgres interactions and migrations
respectively.
Added NATS-Pillow for embedded NATS server. To use for in the future but
also to leverage it's KV for session storage.
Implemented discord OAuth flow for logging in.

Still need to handle NATS superclustering with KV, update README
information, test out the CI of SQLc & Goose migrations, and configure
Fly volumes for NATS JS storage.